Move debugging into the driver to get logging from the first moment
